requirements.txt
Flask==3.0.0
openpyxl==3.1.2
python-dotenv==1.0.0
waitress==2.1.2
azure-cosmos==4.5.1
gunicorn==21.2.0
